feat: install eslint + fix error
